requirements.txt
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.4

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Async task queue
celery>=5.3.0
celery[redis]>=5.3.0

# AI/ML
anthropic>=0.18.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-anthropic>=0.1.0

# Data sources
yfinance>=0.2.36
fredapi>=0.5.1
requests>=2.31.0
aiohttp>=3.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Development
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
